pyproject.toml
[project]
name = "coeffinance"
version = "0.1.0"
description = "Expert system for ratio financial analysis of business."
readme = "README.md"

requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.14",
    "aiosmtplib>=4.0.1",
    "alembic>=1.16.4",
    "alembic-postgresql-enum>=1.8.0",
    "asgi-lifespan>=2.1.0",
    "asyncpg>=0.30.0",
    "cadwyn[standard]>=5.4.4",
    "dishka>=1.6.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.11",
    "fastapi-pagination>=0.14.0",
    "fastapi-users[redis,sqlalchemy]>=14.0.1",
    "fastcrud>=0.15.12",
    "httpx>=0.28.1",
    "pycountry>=24.6.1",
    "pydantic-extra-types>=2.10.2",
    "pydantic-settings>=2.8.1",
    "semver>=3.0.4",
    "sqlakeyset>=2.0.1746777265",
    "sqlmodel>=0.0.24",
    "tenacity>=9.1.2",
    "uvicorn>=0.34.0",
    "zxcvbn>=4.5.0",
]

[dependency-groups]
dev = [
    "asyncpg-stubs>=0.30.1",
    "mypy>=1.15.0",
    "pyright>=1.1.404",
    "ruff>=0.9.10",
    "types-redis>=4.6.0.20241004",
    "types-zxcvbn>=4.5.0.20250809",
]
test = [
    "polyfactory>=2.22.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-icdiff>=0.9",
    "testcontainers[postgres]>=4.13.0",
]

[tool.alembic]
script_location = "%(here)s/src/core/db/migrator"
prepend_sys_path = [
    ".",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
exclude = [
    "src/core/db/migrator/versions",
]

strict = true
# Enables most rules that are not even included in strict mode.
warn_unreachable = true
warn_incomplete_stub = true

plugins = [
    'pydantic.mypy',
]

[[tool.mypy.overrides]]
module = ["testcontainers.*"]
# The lib is de facto typed: https://github.com/testcontainers/testcontainers-python/issues/305.
follow_untyped_imports = true

[tool.pyright]
exclude = [
    "**/.*",
    "**/.*/**",
    "**/__pycache__",
    "src/core/db/migrator/versions",
]

# Separates MyPy ignores, as this type checker does not have a narrowed
# directive such as «mypy: ignore».
enableTypeIgnoreComments = false

typeCheckingMode = "strict"
# Enables most rules that are not even included in strict mode.
deprecateTypingAliases = true
reportCallInDefaultInitializer = "error"
reportImplicitOverride = "error"
reportImportCycles = "error"
reportPropertyTypeMismatch = "error"
reportShadowedImports = "error"
reportUninitializedInstanceVariable = "error"
reportUnnecessaryTypeIgnoreComment = "error"
reportUnreachable = "error"

# Due to the presence of many third-party libraries with a minor lack of typing.
reportUnknownParameterType = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownVariableType = false
reportUnknownMemberType = false

[tool.ruff]
line-length = 80

exclude = [
    "src/core/db/migrator/versions",
]

[tool.ruff.lint]
select = [
    "ALL",
]

ignore = [
    "ARG",
    "D",
    "EM",

    "ANN401",
    "COM812",
    "RUF002",
    "S101",
    "TRY003",
]
